fix: refactor initial

+++ src/main/java/br/com/alelo/consumer/consumerpat/respository/ConsumerRepository.java
@@ -1,12 +1,14 @@
 package br.com.alelo.consumer.consumerpat.respository;
 
-import br.com.alelo.consumer.consumerpat.entity.Consumer;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import java.util.List;
+import br.com.alelo.consumer.consumerpat.entity.Consumer;
 
-public interface ConsumerRepository extends JpaRepository<Consumer, Integer> {
+public interface ConsumerRepository extends JpaRepository<Consumer, Long> {
+	
 
     @Query(nativeQuery = true, value = "select * from Consumer")
     List<Consumer> getAllConsumersList();
